Add files via upload
This was a tool I made to identify specific, preregistered skillsets on CVs for niche recruiters. It is only in a basic stage right now, and would encounter alot of issues, but was a means for me to test my existing knowledge of JS on something relevant to my industry.